AtomsToolbox

Getter functions

AtomsToolbox.distance — Function

distance(system::AbstractSystem, at1, at2)

Get the distance between atoms with indices at1 and at2.

Base.angle — Method

angle(system::AbstractSystem, at1, at2, at3)

Get the angle between vectors connecting atoms with indices at2-at1 and at2-at3.

AtomsToolbox.dihedral — Function

dihedral(system::AbstractSystem, at1, at2, at3, at4)

Get the dihedral angle for the atoms specified by indices at1, at2, at3 and at4.

AtomsToolbox.distance_matrix — Function

distance_matrix(system::AbstractSystem)

Get the distance matrix between all atoms as an NxN matrix where N is the number of atoms in the given system.

AtomsToolbox.scaled_position — Function

scaled_position(sys::AbstractSystem)
scaled_position(atom::Atom, cellmat)

Obtain the scaled positions (fractional coordinates) with respect to the cell matrix for the given system or atom.

AtomsToolbox.cell_parameters — Function

cell_parameters(sys::AbstractSystem)

Obtain the cell lengths and angles as a Vector, [a,b,c,α,β,γ].

AtomsToolbox.cell_lengths — Function

cell_lengths(sys::AbstractSystem)

Obtain the cell lengths as a Vector, [a,b,c].

AtomsToolbox.cell_angles — Function

cell_angles(sys::AbstractSystem)

Obtain the cell angles as a Vector, [α,β,γ].

AtomsToolbox.volume — Function

volume(sys::AbstractSystem)

Obtain the volume of the given system.

AtomsToolbox.connectivity_matrix — Function

connectivity_matrix(
    system;                                                                 
    nlcutoff = natural_cutoffs(system).+0.2,
    retdistmat = false
)

Get the connectivity matrix of the given system. The cutoff is specified using nlcutoff. To return the distance matrix, set retdistmat as true.

AtomsToolbox.connected_components — Function

connected_components(
    system, dists=true;
    nlcutoff = natural_cutoffs(str).+0.2,
    retconnmat = false
)

Get connected components of the given system based on a given cutoff for each of the elements present, nlcutoff which by default is the covalent radii. dists and retconnmat specifies whether to return the distance matrix and connectivity matrix respectively.

AtomsToolbox.pbc_shortest_vectors — Function

pbc_shortest_vectors(
    system::AbstractSystem, at1::Int, at2::Int,
    return_dists::Val{RD}=Val(false), return_vects::Val{RV}=Val(true)
) where {RD, RV}
pbc_shortest_vectors(
    system::AbstractSystem, pos1::T, pos2::T, 
    return_dists::Val{RD}=Val(false), return_vects::Val{RV}=Val(true)
) where {T <: AbstractVector{<: Unitful.Length}, RD, RV}

Obtain the shortest vectors between position vectors, pos1 and pos2, taking into account PBC of the given system.

Functions to build/tranform systems

Base.sort — Method

sort(system::AbstractSystem; by=x->atomic_number(x), rev=false)

Sort an the atoms in an AbstractSystem. The by keyword argument specifies the sorting function and takes as input the atom object. By default, the system is sorted based on its atomic numbers. rev can be used to reverse the order.

AtomsToolbox.supercell — Function

supercell(system::AbstractSystem, supercellvec::Vector{Int}; sorted=false)

Create a supercell of the given system, where supercellvec is the repetitions in each direction. All system and atom properties are copied to the new system. If sorted is true, resultant supercell is sorted based on atomic_symbol of input structure.

AtomsToolbox.wrap — Function

wrap(system::AbstractSystem)

Wrap all the atoms in the given system into the cell and return the wrapped system.

Index

AtomsToolbox.cell_angles

AtomsToolbox.cell_lengths

AtomsToolbox.cell_parameters

AtomsToolbox.connected_components

AtomsToolbox.connectivity_matrix

AtomsToolbox.dihedral

AtomsToolbox.distance

AtomsToolbox.distance_matrix

AtomsToolbox.pbc_shortest_vectors

AtomsToolbox.scaled_position

AtomsToolbox.supercell

AtomsToolbox.volume

AtomsToolbox.wrap

Base.angle

Base.sort